+++ src/docset.rs
@@ -82,7 +82,7 @@ pub trait DocSet: Send {
 
     /// Returns a best-effort hint of the
     /// length of the docset.
-    fn size_hint(&self) -> u32;
+    fn size_hint(&self) -> u64;
 
     /// Returns the number documents matching.
     /// Calling this method consumes the `DocSet`.
@@ -127,7 +127,7 @@ impl<'a> DocSet for &'a mut dyn DocSet {
         (**self).doc()
     }
 
-    fn size_hint(&self) -> u32 {
+    fn size_hint(&self) -> u64 {
         (**self).size_hint()
     }
 
@@ -161,7 +161,7 @@ impl<TDocSet: DocSet + ?Sized> DocSet for Box<TDocSet> {
         unboxed.doc()
     }
 
-    fn size_hint(&self) -> u32 {
+    fn size_hint(&self) -> u64 {
         let unboxed: &TDocSet = self.borrow();
         unboxed.size_hint()
     }
