+++ src/postings/mod.rs
@@ -519,7 +519,7 @@ pub mod tests {
             self.0.doc()
         }
 
-        fn size_hint(&self) -> u32 {
+        fn size_hint(&self) -> u64 {
             self.0.size_hint()
         }
     }
