+++ src/query/phrase_prefix_query/phrase_prefix_scorer.rs
@@ -55,7 +55,7 @@ impl<TPostings: Postings> DocSet for PhraseKind<TPostings> {
         }
     }
 
-    fn size_hint(&self) -> u32 {
+    fn size_hint(&self) -> u64 {
         match self {
             PhraseKind::SinglePrefix { postings, .. } => postings.size_hint(),
             PhraseKind::MultiPrefix(postings) => postings.size_hint(),
@@ -193,7 +193,7 @@ impl<TPostings: Postings> DocSet for PhrasePrefixScorer<TPostings> {
         self.phrase_scorer.doc()
     }
 
-    fn size_hint(&self) -> u32 {
+    fn size_hint(&self) -> u64 {
         self.phrase_scorer.size_hint()
     }
 }
