+++ sstable/src/delta.rs
@@ -41,11 +41,11 @@ where
         self.block_len = block_len
     }
 
-    pub fn flush_block(&mut self) -> io::Result<Option<Range<usize>>> {
+    pub fn flush_block(&mut self) -> io::Result<Option<Range<u64>>> {
         if self.block.is_empty() {
             return Ok(None);
         }
-        let start_offset = self.write.written_bytes() as usize;
+        let start_offset = self.write.written_bytes();
 
         let buffer: &mut Vec<u8> = &mut self.stateless_buffer;
         self.value_writer.serialize_block(buffer);
@@ -81,7 +81,7 @@ where
             self.write.write_all(&self.block[..])?;
         }
 
-        let end_offset = self.write.written_bytes() as usize;
+        let end_offset = self.write.written_bytes();
         self.block.clear();
         buffer.clear();
         Ok(Some(start_offset..end_offset))
@@ -110,7 +110,7 @@ where
         self.value_writer.write(value);
     }
 
-    pub fn flush_block_if_required(&mut self) -> io::Result<Option<Range<usize>>> {
+    pub fn flush_block_if_required(&mut self) -> io::Result<Option<Range<u64>>> {
         if self.block.len() > self.block_len {
             return self.flush_block();
         }
